Added -fno-stack-protector flag, made terminal frontend compile
The change to the flags was made a long while back to get something
working.  I don't recall the details.

The frontend had a dependency on a generated config.h file that
would have been generated if I had generated the config files like
the library intended.  However, a couple of quick changes made it
easy to compile the frontend.  I needed it to quickly test udis86's
compatibility with some instructions.