Fixed decoding of certain vpsllq and vpslld instructions
I found that vpsllq and vpslld instructions that were encoded with
an immediate byte were not being decoded correctly.  After a lot of
research into the instructions, I discovered that the operands for
these forms of these instructions were not correctly defined in
optable.xml.  I have therefore provided the correct operands for
these instructions, and now udis86 outputs the same result as gdb,
objdump and capstone.